requirements.txt
opencv-python
torch
torchvision
seaborn
pillow
gradio